# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Shop API
  version: 0.1.0
  description: |
    Черновой контракт API интернет-магазина кофе и кофемашин (Epic 0).
    Файл — стартовая точка для параллельной разработки backend/frontend.
    После реализации эндпоинтов источником истины станет схема из
    **drf-spectacular** (`GET /api/schema/`), а этот YAML обновляется по мере расхождений.

    ## Конвенции

    ### Пагинация
    List-эндпоинты используют query-параметры **`page`** (с 1) и **`page_size`** (по умолчанию 20, max 100).
    Ответ обёрнут в объект:

    ```json
    {
      "count": 42,
      "next": "https://api.example/...?page=3",
      "previous": "https://api.example/...?page=1",
      "results": []
    }
    ```

    ### Ошибки
    Единый формат тела ошибки:

    ```json
    {
      "detail": "Human-readable message",
      "code": "machine_readable_code"
    }
    ```

    Для ошибок валидации полей допускается расширение:

    ```json
    {
      "detail": "Invalid input",
      "code": "validation_error",
      "errors": { "email": ["Enter a valid email."] }
    }
    ```

    ### Цены
    Денежные суммы — **строка** с десятичным числом (например `"1299.00"`), валюта всегда **RUB**.
    Отдельного поля валюты в MVP нет (мультивалютность не поддерживается).

    ### Дата и время
    Все даты/время — **ISO 8601** в UTC, например `2026-07-25T12:00:00Z`.
servers:
  - url: http://127.0.0.1:8000
    description: Local development
...
paths:
...
  /api/auth/refresh/:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh:
                  type: string
                  description: Refresh JWT
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                required: [access]
                properties:
                  access:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
          description: Human-readable message
        code:
          type: string
          description: Machine-readable error code
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Optional field-level validation errors
...
security:
  - bearerAuth: []